Fix #162, reviewed ACU status message generation.
Now the various status frames that compose the whole ACU status message are kept in memory as `bytearray` objects, which acts as mutable strings. This means that the various status frames do not have to be composed every time the message is about to be sent. The statuses gets updated as soon as a property is modified, minimizing the time needed to send the message to connected clients.