Fix #200, improved AS simulator
Now the positioning of each USD is handled by a single thread for each line, instead of having a thread for each USD, which used to load too much the machine (more than 1000 concurrent threads). The whole positioning strategy has been refactored, along with some other fixes and improvements. Tests have been updated as well, in order to grant as much coverage as possible. The documentation though has to be updated ASAP in order for this branch to be merged.