Fix #217, fix #218, see changelog below.
Fix #217, now the `discos-simulator` command is able to start all the
simulators instead of a single one per command. In order to do so, you
just have to issue the following command:
`discos-simulator start`.
Fix #218, in order to enable #217, a list of all simulators implemented
in the package had to be defined or automatically retrieved as a
function. The latter was the chosen approach, a `list_simulator`
function was implemented inside the `utils` library.

This commit brings also some major improvements regarding the
`MultiTypeSystem` class. The way the desired system gets chosen for this
kind of simulator is now cleaner and definitely more pythonic (the
desired system type is now chosen as a keyworded argument).