Fix #203, added the receiver simulator
The current receiver simulator only implements the CNR protocol. In
order to make it work as a proper receiver, the `Slave` class should be
specialized in order to answer to some interrogation commands in the
correct way. The next step should be to analyze the `ReceiversLibrary`
code and documentation, in order to properly replicate the behavior of
the receiver's slave board(s).