Fix #203, added the receiver simulator (#230)
* Fix #203, added the receiver simulator

The current receiver simulator only implements the CNR protocol. In
order to make it work as a proper receiver, the `Slave` class should be
specialized in order to answer to some interrogation commands in the
correct way. The next step should be to analyze the `ReceiversLibrary`
code and documentation, in order to properly replicate the behavior of
the receiver's slave board(s).

* Committing a few additions

* Dewar tested, improved receivers coverage

* Fix #203, implemented LNA stage values reading

* Fix #203, improved implementation for receiver simulator

* Fix #203, improved commands with parameters

* Fixed a wrong behavior of local oscillator simulator