Fix #312, removed hardcoded ports from tests
Hardcoded ports were removed from test_server.py and they were replaced
with a python generator that checks if a given port is available, then
use that port for the next test. This allows multiple instances of the
same test file to be executed concurrently without clashing issues.
In order to enable this feature a minor fix was introduced regarding the
logging style, now the logger also prints to file the PID of the process
logging some information. The PID is then used when testing and reading
from the default logfile in order for each test process to read only its
lines without issues.